nmk/nmk16.cpp: upgraded interrupts trigger system based on PROM contents * added info about screen resolutions and timings * merged powerins.cpp into nmk16.cpp driver to make use the new interrupt system
