+++ src/mame/nmk/nmk16.h
@@ -77,6 +77,13 @@ public:
 	void mustangb3(machine_config &config);
 	void twinactn(machine_config &config);
 	void vandykeb(machine_config &config);
+	void powerins(machine_config &config);
+	void powerinsj(machine_config &config);
+	void powerinspu(machine_config &config);
+	void powerinspj(machine_config &config);
+	void powerinsa(machine_config &config);
+	void powerinsb(machine_config &config);
+	void powerinsc(machine_config &config);
 
 	void init_nmk();
 	void init_tharrier();
@@ -87,10 +94,15 @@ public:
 	void init_banked_audiocpu();
 	void init_gunnailb();
 	void init_bjtwin();
+	void init_powerinsa();
 	void init_acrobatmbl();
 
 protected:
+	virtual void machine_start() override;
+	virtual void machine_reset() override;
+
 	TIMER_DEVICE_CALLBACK_MEMBER(nmk16_scanline);
+	TIMER_DEVICE_CALLBACK_MEMBER(nmk16_hacky_scanline);
 	u32 screen_update_macross(screen_device &screen, bitmap_ind16 &bitmap, const rectangle &cliprect);
 
 	void txvideoram_w(offs_t offset, u16 data, u16 mem_mask = ~0);
@@ -98,7 +110,12 @@ protected:
 	void vandyke_flipscreen_w(u8 data);
 	void tilebank_w(u8 data);
 
-	void macross2_sound_bank_w(u8 data);
+	void macross2_sound_reset_w(u16 data);
+	void macross2_audiobank_w(u8 data);
+	void ssmissin_okibank_w(u8 data);
+	void powerinsa_okibank_w(u8 data);
+	template<unsigned Chip> void tharrier_okibank_w(u8 data);
+	u8 powerins_bootleg_fake_ym2203_r();
 
 	required_device<cpu_device> m_maincpu;
 	optional_device<cpu_device> m_audiocpu;
@@ -138,15 +155,13 @@ protected:
 	u8 m_scroll[2][4]{};
 	u16 m_vscroll[4]{};
 	int m_prot_count = 0;
+	u8 m_interrupt_trigger;
 
 	void mainram_strange_w(offs_t offset, u16 data/*, u16 mem_mask = ~0*/);
 	u16 mainram_swapped_r(offs_t offset);
 	void mainram_swapped_w(offs_t offset, u16 data, u16 mem_mask = ~0);
-	void ssmissin_soundbank_w(u8 data);
 	void tharrier_mcu_control_w(u16 data);
 	u16 tharrier_mcu_r(offs_t offset, u16 mem_mask = ~0);
-	void macross2_sound_reset_w(u16 data);
-	template<unsigned Chip> void tharrier_oki_bankswitch_w(u8 data);
 	u16 vandykeb_r();
 	u16 tdragonb_prot_r();
 	template<unsigned Layer> void bgvideoram_w(offs_t offset, u16 data, u16 mem_mask = ~0);
@@ -161,27 +176,34 @@ protected:
 	void nmk004_x0016_w(u16 data);
 	void nmk004_bioship_x0016_w(u16 data);
 
+	void set_interrupt_timing(machine_config &config);
 	void set_hacky_interrupt_timing(machine_config &config);
-	void set_hacky_screen_lowres(machine_config &config);
-	void set_hacky_screen_hires(machine_config &config);
+	void set_screen_lowres(machine_config &config);
+	void set_screen_midres(machine_config &config);
+	void set_screen_hires(machine_config &config);
 
 	TILEMAP_MAPPER_MEMBER(tilemap_scan_pages);
 	template<unsigned Layer, unsigned Gfx> TILE_GET_INFO_MEMBER(common_get_bg_tile_info);
 	TILE_GET_INFO_MEMBER(common_get_tx_tile_info);
 	TILE_GET_INFO_MEMBER(bioship_get_bg_tile_info);
 	TILE_GET_INFO_MEMBER(bjtwin_get_bg_tile_info);
+	TILE_GET_INFO_MEMBER(powerins_get_bg_tile_info);
 	DECLARE_VIDEO_START(macross);
 	DECLARE_VIDEO_START(bioship);
 	DECLARE_VIDEO_START(strahl);
 	DECLARE_VIDEO_START(macross2);
 	DECLARE_VIDEO_START(gunnail);
 	DECLARE_VIDEO_START(bjtwin);
+	DECLARE_VIDEO_START(powerins);
 	void get_colour_4bit(u32 &colour, u32 &pri_mask);
 	void get_colour_5bit(u32 &colour, u32 &pri_mask);
+	void get_colour_6bit(u32 &colour, u32 &pri_mask);
 	void get_sprite_flip(u16 attr, int &flipx, int &flipy, int &code);
+	void get_flip_extcode_powerins(u16 attr, int &flipx, int &flipy, int &code);
 	u32 screen_update_tharrier(screen_device &screen, bitmap_ind16 &bitmap, const rectangle &cliprect);
 	u32 screen_update_strahl(screen_device &screen, bitmap_ind16 &bitmap, const rectangle &cliprect);
 	u32 screen_update_bjtwin(screen_device &screen, bitmap_ind16 &bitmap, const rectangle &cliprect);
+	void screen_vblank_powerins_bootleg(int state);
 	TIMER_CALLBACK_MEMBER(dma_callback);
 	TIMER_DEVICE_CALLBACK_MEMBER(manybloc_scanline);
 	void video_init();
@@ -216,6 +238,11 @@ protected:
 	void mustangb3_sound_map(address_map &map);
 	void oki1_map(address_map &map);
 	void oki2_map(address_map &map);
+	void powerins_map(address_map &map);
+	void powerins_sound_map(address_map &map);
+	void powerins_bootleg_audio_io_map(address_map &map);
+	void powerinsa_map(address_map &map);
+	void powerinsa_oki_map(address_map &map);
 	void raphero_map(address_map &map);
 	void raphero_sound_mem_map(address_map &map);
 	void ssmissin_map(address_map &map);
